% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_long_surv_jm.R
\name{cond_surv_prob}
\alias{cond_surv_prob}
\title{Compute the conditional probability given that subjects survived upto time t}
\usage{
cond_surv_prob(Obj, lm_time)
}
\arguments{
\item{Obj}{object returned by predict.supFTSVD_JM}

\item{lm_time}{landmark time point where the subjects were alive}
}
\description{
Compute the conditional probability given that subjects survived upto time t
}
