% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_long_surv_jm.R
\name{gfold_supFTSVD_JM}
\alias{gfold_supFTSVD_JM}
\title{It splits data into g-fold and fits supFTSVD-JM, separately treating each
fold as test data.}
\usage{
gfold_supFTSVD_JM(
  datlist,
  response,
  interval = NULL,
  r = 1,
  resolution = 100,
  CVPhi = FALSE,
  K = 5,
  cvT = 5,
  smooth = 1e-08,
  maxiter = 20,
  epsilon = 1e-04,
  KInd = NULL,
  surv_time,
  censor_status,
  survX = NULL,
  mc_ss = 1000,
  rsvd_seed = 100,
  conv_criteria = "par",
  rank_fold = 5,
  rfolds_seed = 50,
  stratum = NULL,
  par_computing = TRUE,
  n_core = 5,
  scale = TRUE,
  constant_hazard = TRUE,
  change_points = NULL
)
}
\arguments{
\item{datlist}{a list with n elements, each is a matrix of dimension p+1
times m_i. The first row represents time points where measurements were
obtained. Next p rows represent the function values observed at the time
points}

\item{response}{a matrix of dimension n by q contains the design matrix for
supervision of the subject loadings}

\item{interval}{represents the domain of the function. The default is NULL.}

\item{r}{rank of the supFTSVD model}

\item{resolution}{grid size on which the singular functions will be estimated}

\item{CVPhi}{a logical scalar representing whether cross-validation (CV) will be
performed for determining the smoothness parameters involved with singular
functions}

\item{K}{number of folds for the CV step}

\item{cvT}{number of initial iterations at which CV will be performed}

\item{smooth}{Smoothing parameter for RKHS norm when CVPhi is FALSE. With
CVPhi=TRUE, a vector of numeric values for smoothness parameters for grid
search}

\item{maxiter}{Maximum number of iteration. Default: 20.}

\item{epsilon}{Convergence threshold for objective function value}

\item{KInd}{is a vector indices to construct the folds for CV. The default is
NULL that indicates that folds will be constructed randomly inside}

\item{rsvd_seed}{supFTSVD uses random svd for computational advantage while
setting inial values for feature loading vector. This seed number will be used
before setting the initial values. It confirms the reproducibility of your results.}

\item{conv_criteria}{the default is par which sets the threshold on the
difference between updated and current estimate at M step to stop the iteration.
It is computationally expensive; thus we recommend to use "cond_lik" instead of "par".
The result is not sensitive.}

\item{rank_fold}{number of folds to be used in the cross-validation}

\item{rfolds_seed}{seed number to be used in random splitting of the data}
}
\value{
a list with following objects:
\itemize{
\item fold_res : detailed results for different folds
\item rank : rank of the fitted model
}
}
\description{
It splits data into g-fold and fits supFTSVD-JM, separately treating each
fold as test data.
}
