% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_long_surv_jm.R
\name{r2_supFTSVD_JM}
\alias{r2_supFTSVD_JM}
\title{It calculates R2, square of Pearson correlation coefficient, and R2 if a
linear model would have been fitted using the fitted components.}
\usage{
r2_supFTSVD_JM(obj, obs_data)
}
\arguments{
\item{obj}{fitted object obtained by supFTSVD_JM}

\item{obs_data}{observed data used in the fitted object}
}
\description{
It calculates R2, square of Pearson correlation coefficient, and R2 if a
linear model would have been fitted using the fitted components.
}
