% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_long_surv_jm.R
\name{summary_coxph_supFTSVD_JM}
\alias{summary_coxph_supFTSVD_JM}
\title{Provides similar summary as survival::summary.coxph}
\usage{
summary_coxph_supFTSVD_JM(object)
}
\arguments{
\item{object}{takes the fitted model object obtained by fitting supFTSVD_JM}
}
\value{
it returns a data frame that contains estimated coefficient, hazard
ratio, and associated Wald-type test for the null hypothesis of zero
}
\description{
Provides similar summary as survival::summary.coxph
}
\examples{
load(file = "examples/fit_model.RData")
summary_coxph_supFTSVD_JM(fit_model)
}
